+++ packages/apollo-codegen-swift/src/__tests__/codeGeneration.ts
@@ -222,7 +222,7 @@ describe("Swift code generation", () => {
           responseKey: "response_key",
           propertyName: "propertyName",
           type: GraphQLString
-        })
+        }).source
       ).toBe('"response_key": propertyName');
     });
 
@@ -232,7 +232,7 @@ describe("Swift code generation", () => {
           responseKey: "response_key",
           propertyName: "propertyName",
           type: new GraphQLNonNull(GraphQLString)
-        })
+        }).source
       ).toBe('"response_key": propertyName');
     });
 
@@ -242,7 +242,7 @@ describe("Swift code generation", () => {
           responseKey: "response_key",
           propertyName: "propertyName",
           type: new GraphQLList(GraphQLString)
-        })
+        }).source
       ).toBe('"response_key": propertyName');
     });
 
@@ -252,7 +252,7 @@ describe("Swift code generation", () => {
           responseKey: "response_key",
           propertyName: "propertyName",
           type: new GraphQLList(new GraphQLNonNull(GraphQLString))
-        })
+        }).source
       ).toBe('"response_key": propertyName');
     });
 
@@ -262,7 +262,7 @@ describe("Swift code generation", () => {
           responseKey: "response_key",
           propertyName: "propertyName",
           type: new GraphQLNonNull(new GraphQLList(GraphQLString))
-        })
+        }).source
       ).toBe('"response_key": propertyName');
     });
 
@@ -274,7 +274,7 @@ describe("Swift code generation", () => {
           type: new GraphQLNonNull(
             new GraphQLList(new GraphQLNonNull(GraphQLString))
           )
-        })
+        }).source
       ).toBe('"response_key": propertyName');
     });
 
@@ -284,7 +284,7 @@ describe("Swift code generation", () => {
           responseKey: "response_key",
           propertyName: "propertyName",
           type: schema.getType("Droid")
-        })
+        }).source
       ).toBe(
         '"response_key": propertyName.flatMap { (value: Droid) -> ResultMap in value.resultMap }'
       );
@@ -296,7 +296,7 @@ describe("Swift code generation", () => {
           responseKey: "response_key",
           propertyName: "propertyName",
           type: new GraphQLNonNull(schema.getType("Droid"))
-        })
+        }).source
       ).toBe('"response_key": propertyName.resultMap');
     });
 
@@ -306,7 +306,7 @@ describe("Swift code generation", () => {
           responseKey: "response_key",
           propertyName: "propertyName",
           type: new GraphQLList(schema.getType("Droid"))
-        })
+        }).source
       ).toBe(
         '"response_key": propertyName.flatMap { (value: [Droid?]) -> [ResultMap?] in value.map { (value: Droid?) -> ResultMap? in value.flatMap { (value: Droid) -> ResultMap in value.resultMap } } }'
       );
@@ -318,7 +318,7 @@ describe("Swift code generation", () => {
           responseKey: "response_key",
           propertyName: "propertyName",
           type: new GraphQLList(new GraphQLNonNull(schema.getType("Droid")))
-        })
+        }).source
       ).toBe(
         '"response_key": propertyName.flatMap { (value: [Droid]) -> [ResultMap] in value.map { (value: Droid) -> ResultMap in value.resultMap } }'
       );
@@ -330,7 +330,7 @@ describe("Swift code generation", () => {
           responseKey: "response_key",
           propertyName: "propertyName",
           type: new GraphQLNonNull(new GraphQLList(schema.getType("Droid")))
-        })
+        }).source
       ).toBe(
         '"response_key": propertyName.map { (value: Droid?) -> ResultMap? in value.flatMap { (value: Droid) -> ResultMap in value.resultMap } }'
       );
@@ -344,7 +344,7 @@ describe("Swift code generation", () => {
           type: new GraphQLNonNull(
             new GraphQLList(new GraphQLNonNull(schema.getType("Droid")))
           )
-        })
+        }).source
       ).toBe(
         '"response_key": propertyName.map { (value: Droid) -> ResultMap in value.resultMap }'
       );
@@ -440,6 +440,9 @@ describe("Swift code generation", () => {
         query Hero {
           hero {
             private: name
+            self: friends {
+              id
+            }
           }
         }
       `);
@@ -661,7 +664,7 @@ describe("Swift code generation", () => {
         .selectionSet.selections[0] as Field).args as Argument[];
       const dictionaryLiteral = generator.helpers.dictionaryLiteralForFieldArguments(
         fieldArguments
-      );
+      ).source;
 
       expect(dictionaryLiteral).toBe(
         '["episode": "JEDI", "review": ["stars": 2, "commentary": GraphQLVariable("commentary"), "favorite_color": ["red": GraphQLVariable("red"), "blue": 100, "green": 50]]]'