+++ packages/apollo/src/commands/service/check.ts
@@ -287,7 +287,8 @@ export default class ServiceCheck extends ProjectCommand {
     }),
     localSchemaFile: flags.string({
       description:
-        "Path to your local GraphQL schema file (introspection result or SDL)"
+        "Path to your local GraphQL schema file (introspection result or SDL)",
+      multiple: true
     }),
     markdown: flags.boolean({
       description: "Output result in markdown.",